+++ src/phoneField/index.tsx
@@ -5,7 +5,7 @@ import isEmpty from 'lodash.isempty'
 import isEqual from 'lodash.isequal'
 import SelectField from 'selectField'
 import TextField from 'textField'
-import countryTelData from 'country-telephone-data'
+import { allCountries, allCountryCodes } from 'country-telephone-data'
 import style from 'phoneField/style'
 
 enum FIELDS {
@@ -17,75 +17,76 @@ interface PhoneFieldOnChangeParameters {
   value: {
     [FIELDS.PHONENUMBER]: string
     [FIELDS.PHONEPREFIX]: string
+    phoneRegionList: string[]
     completePhoneNumber: string
   }
 }
 interface PhoneFieldProps {
   readonly name: string
+  readonly onChange: (obj: PhoneFieldOnChangeParameters) => void
   readonly id?: string
   readonly className?: string
-  readonly defaultPrefix?: string
-  readonly prefixLabel?: string
-  readonly title?: string
   readonly ariaLabelledBy?: string
-  readonly placeholder?: string
-  readonly defaultValue?: string
-  readonly countriesWhiteList?: string[]
-  readonly customCountriesName?: customCountriesName
+  readonly selectFieldLabel?: string
+  readonly textFieldTitle?: string
+  readonly textFieldPlaceholder?: string
+  readonly defaultRegionValue?: string
+  readonly defaultPhoneValue?: string
+  readonly countryWhitelist?: string[]
+  readonly customCountryNames?: customCountryNames
   readonly isInline?: boolean
-  readonly onChange?: (obj: PhoneFieldOnChangeParameters) => void
 }
 interface PhoneFieldState {
-  countriesData: mappedCountriesPhoneData[]
-  countriesWhiteList: string[]
+  countryData: mappedCountryPhoneData[]
+  countryWhitelist: string[]
+  phoneRegionList: string[]
   completePhoneNumber: string
   // Make the type of [key: string]: string etc... assignable for setState()
   [key: string]: any
 }
 
-const allCountriesPhoneData: allCountriesPhoneData[] = countryTelData.allCountries
+const allCountryPhoneData: allCountryPhoneData[] = allCountries
 
 /* Format and keep only used data */
-const formattedCountriesPhoneData: formattedCountriesPhoneData[] = allCountriesPhoneData.map(
+const formattedCountryPhoneData: formattedCountryPhoneData[] = allCountryPhoneData.map(
   countryData => ({
     name: countryData.name,
     iso2: countryData.iso2.toUpperCase(),
-    dialCode: countryData.dialCode,
+    dialCode: `+${countryData.dialCode}`,
   }),
 )
-
 /**
  * Filter countries data according to whitelist
- * @param {Array} countriesList in ISO2 format ex: ['FR', 'ES']
+ * @param {Array} countryList in ISO2 format ex: ['FR', 'ES']
  * @return {formattedCountriesPhoneData}
  */
-const filterIso2 = (countriesList: string[]) =>
-  formattedCountriesPhoneData.filter(country => countriesList.includes(country.iso2))
+const filterIso2 = (countryList: string[]) =>
+  formattedCountryPhoneData.filter(country => countryList.includes(country.iso2))
 
 /**
  * Return country data according to ISO2 defaultValue
  * @param {string} countryDefault ISO2 format ex: 'FR'
  * @return {formattedCountriesPhoneData}
  */
-const findIso2 = (countryDefault: string): formattedCountriesPhoneData =>
-  formattedCountriesPhoneData.find(country => country.iso2 === countryDefault)
+const findIso2 = (countryDefault: string): formattedCountryPhoneData =>
+  formattedCountryPhoneData.find(country => country.iso2 === countryDefault)
 
 const mapCountriesPhoneData = (
-  countriesData: formattedCountriesPhoneData[],
-  countriesName: customCountriesName,
-): mappedCountriesPhoneData[] => {
-  return countriesData.map(countryData => ({
-    value: `+${countryData.dialCode}`,
-    label: `${
-      !isEmpty(countriesName[countryData.iso2]) ? countriesName[countryData.iso2] : countryData.name
-    } +${countryData.dialCode}`,
+  countryData: formattedCountryPhoneData[],
+  countryNames: customCountryNames,
+): mappedCountryPhoneData[] => {
+  return countryData.map(data => ({
+    value: data.dialCode,
+    label: `${!isEmpty(countryNames[data.iso2]) ? countryNames[data.iso2] : data.name} ${
+      data.dialCode
+    }`,
   }))
 }
 
 /* Alphabetically sorted */
 const sortCountriesPhoneData = (
-  mappedCountryPhoneData: mappedCountriesPhoneData[],
-): mappedCountriesPhoneData[] =>
+  mappedCountryPhoneData: mappedCountryPhoneData[],
+): mappedCountryPhoneData[] =>
   mappedCountryPhoneData.sort((a, b) => {
     if (a.label < b.label) return -1
     if (a.label > b.label) return 1
@@ -98,63 +99,61 @@ const iso2toDialCode = (iso2Value: string) => {
   }
   const countryPhoneData = findIso2(iso2Value)
   if (countryPhoneData) {
-    return `+${countryPhoneData.dialCode}`
+    return countryPhoneData.dialCode
   }
   return ''
 }
 
 export default class PhoneField extends PureComponent<PhoneFieldProps, PhoneFieldState> {
   static defaultProps: Partial<PhoneFieldProps> = {
-    defaultPrefix: '',
-    defaultValue: '',
-    countriesWhiteList: [],
-    customCountriesName: {},
+    defaultRegionValue: '',
+    defaultPhoneValue: '',
+    countryWhitelist: [],
+    customCountryNames: {},
     isInline: true,
   }
 
   state = {
-    countriesData: PhoneField.countriesData(
-      this.props.countriesWhiteList,
-      this.props.customCountriesName,
-    ),
-    countriesWhiteList: this.props.countriesWhiteList,
-    [FIELDS.PHONENUMBER]: this.props.defaultValue,
-    [FIELDS.PHONEPREFIX]: iso2toDialCode(this.props.defaultPrefix),
-    completePhoneNumber: iso2toDialCode(this.props.defaultPrefix) + this.props.defaultValue,
+    countryData: PhoneField.countryData(this.props.countryWhitelist, this.props.customCountryNames),
+    countryWhitelist: this.props.countryWhitelist,
+    [FIELDS.PHONENUMBER]: this.props.defaultPhoneValue,
+    [FIELDS.PHONEPREFIX]: iso2toDialCode(this.props.defaultRegionValue),
+    phoneRegionList: !isEmpty(this.props.defaultRegionValue) ? [this.props.defaultRegionValue] : [],
+    completePhoneNumber:
+      iso2toDialCode(this.props.defaultRegionValue) + this.props.defaultPhoneValue,
   }
 
   onChange = ({ name, value }: OnChangeParameters) => {
     this.setState({ [name]: value }, () => {
+      const iso2Codes = allCountryCodes[Number(this.state[FIELDS.PHONEPREFIX])]
       this.props.onChange({
         name: this.props.name,
         value: {
           [FIELDS.PHONENUMBER]: this.state[FIELDS.PHONENUMBER],
           [FIELDS.PHONEPREFIX]: this.state[FIELDS.PHONEPREFIX],
+          phoneRegionList: !isEmpty(iso2Codes) ? iso2Codes : [],
           completePhoneNumber: this.state[FIELDS.PHONEPREFIX] + this.state[FIELDS.PHONENUMBER],
         },
       })
     })
   }
 
-  static countriesData(whitelist: string[], countriesName: customCountriesName) {
+  static countryData(whitelist: string[], countryNames: customCountryNames) {
     if (!isEmpty(whitelist)) {
       const whiteListCountriesPhoneData = filterIso2(whitelist)
-      const whitelistMapped = mapCountriesPhoneData(whiteListCountriesPhoneData, countriesName)
+      const whitelistMapped = mapCountriesPhoneData(whiteListCountriesPhoneData, countryNames)
 
       return sortCountriesPhoneData(whitelistMapped)
     }
-    const countriesDatasMapped = mapCountriesPhoneData(formattedCountriesPhoneData, countriesName)
+    const countriesDatasMapped = mapCountriesPhoneData(formattedCountryPhoneData, countryNames)
     return sortCountriesPhoneData(countriesDatasMapped)
   }
 
   static getDerivedStateFromProps(props: PhoneFieldProps, state: PhoneFieldState) {
-    if (!isEqual(props.countriesWhiteList, state.countriesWhiteList)) {
+    if (!isEqual(props.countryWhitelist, state.countryWhitelist)) {
       return {
-        countriesData: PhoneField.countriesData(
-          props.countriesWhiteList,
-          props.customCountriesName,
-        ),
-        countriesWhiteList: props.countriesWhiteList,
+        countryData: PhoneField.countryData(props.countryWhitelist, props.customCountryNames),
+        countryWhitelist: props.countryWhitelist,
       }
     }
 
@@ -162,7 +161,14 @@ export default class PhoneField extends PureComponent<PhoneFieldProps, PhoneFiel
   }
 
   render() {
-    const { id, prefixLabel, title, ariaLabelledBy, placeholder, defaultValue } = this.props
+    const {
+      id,
+      selectFieldLabel,
+      textFieldTitle,
+      ariaLabelledBy,
+      textFieldPlaceholder,
+      defaultPhoneValue,
+    } = this.props
 
     const baseClassName = cc([prefix({ phoneField: true })])
 
@@ -175,18 +181,18 @@ export default class PhoneField extends PureComponent<PhoneFieldProps, PhoneFiel
       <div id={id} className={classNames} aria-labelledby={ariaLabelledBy}>
         <SelectField
           name={FIELDS.PHONEPREFIX}
-          options={this.state.countriesData}
-          ariaLabel={prefixLabel}
+          options={this.state.countryData}
+          ariaLabel={selectFieldLabel}
           defaultValue={this.state[FIELDS.PHONEPREFIX]}
           onChange={this.onChange}
         />
         <TextField
           type="tel"
           name={FIELDS.PHONENUMBER}
-          placeholder={placeholder}
-          defaultValue={defaultValue}
+          placeholder={textFieldPlaceholder}
+          defaultValue={defaultPhoneValue}
           onChange={this.onChange}
-          title={title}
+          title={textFieldTitle}
         />
         <style jsx global>
           {style}